requirements.txt
streamlit
pandas
groq
python-dotenv
pytest
jinja2
tabulate
